@@ -30,102 +30,103 @@
   },
   "homepage": "https://github.com/hbz/oerworldmap-ui#readme",
   "engines": {
-    "node": ">=8.0.0"
+    "node": ">=12.16.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.5.5",
-    "@babel/preset-env": "^7.5.5",
-    "@babel/preset-react": "^7.0.0",
+    "@babel/core": "^7.8.4",
+    "@babel/preset-env": "^7.8.4",
+    "@babel/preset-react": "^7.8.3",
     "babel-core": "^7.0.0-bridge.0",
-    "babel-eslint": "^10.0.1",
-    "babel-jest": "^24.8.0",
+    "babel-eslint": "^10.0.3",
+    "babel-jest": "^25.1.0",
     "babel-loader": "^8.0.6",
-    "copy-webpack-plugin": "^5.0.4",
-    "css-loader": "^3.1.0",
+    "copy-webpack-plugin": "^5.1.1",
+    "css-loader": "^3.4.2",
     "cssnano": "^4.1.8",
-    "enzyme": "^3.10.0",
-    "enzyme-adapter-react-16": "^1.14.0",
-    "eslint": "^6.1.0",
+    "enzyme": "^3.11.0",
+    "enzyme-adapter-react-16": "^1.15.2",
+    "eslint": "^6.8.0",
     "eslint-config-airbnb": "^17.1.1",
-    "eslint-config-stylelint": "^11.0.0",
-    "eslint-plugin-import": "^2.18.2",
+    "eslint-config-stylelint": "^11.1.0",
+    "eslint-plugin-import": "^2.20.1",
     "eslint-plugin-jsx-a11y": "^6.2.3",
-    "eslint-plugin-react": "^7.14.3",
-    "file-loader": "^4.1.0",
+    "eslint-plugin-react": "^7.18.3",
+    "file-loader": "^5.1.0",
     "identity-obj-proxy": "^3.0.0",
-    "jest": "^24.8.0",
-    "jsdom": "^15.1.1",
-    "mini-css-extract-plugin": "^0.8.0",
+    "jest": "^25.1.0",
+    "mini-css-extract-plugin": "^0.9.0",
     "optimize-css-assets-webpack-plugin": "^5.0.3",
-    "postcss": "^7.0.17",
+    "postcss": "^7.0.27",
     "postcss-loader": "^3.0.0",
-    "postcss-safe-parser": "^4.0.1",
+    "postcss-safe-parser": "^4.0.2",
     "regenerator-runtime": "^0.13.3",
-    "style-loader": "^0.23.1",
-    "stylelint": "^10.1.0",
-    "stylelint-config-standard": "^18.3.0",
-    "webpack-cli": "^3.3.6",
-    "webpack-merge": "^4.2.1"
+    "style-loader": "^1.1.3",
+    "stylelint": "^13.2.0",
+    "stylelint-config-standard": "^20.0.0",
+    "webpack-cli": "^3.3.11",
+    "webpack-merge": "^4.2.2"
   },
   "dependencies": {
-    "@babel/polyfill": "^7.4.4",
-    "@babel/register": "^7.5.5",
-    "@storybook/addon-a11y": "^5.0.11",
-    "@storybook/addon-actions": "^5.0.11",
-    "@storybook/addon-knobs": "^5.0.11",
-    "@storybook/addon-notes": "^5.0.11",
-    "@storybook/addon-storysource": "^5.0.11",
-    "@storybook/react": "^5.0.11",
+    "@appbaseio/reactivesearch": "^3.4.1",
+    "@babel/polyfill": "^7.8.3",
+    "@babel/register": "^7.8.3",
+    "@storybook/addon-a11y": "^5.3.14",
+    "@storybook/addon-actions": "^5.3.14",
+    "@storybook/addon-knobs": "^5.3.14",
+    "@storybook/addon-notes": "^5.3.14",
+    "@storybook/addon-storysource": "^5.3.14",
+    "@storybook/react": "^5.3.14",
     "@turf/bbox": "^6.0.1",
     "@turf/center": "^6.0.1",
     "@turf/helpers": "^6.1.4",
     "@turf/invariant": "^6.1.2",
-    "ajv": "^6.10.2",
+    "ajv": "^6.12.0",
     "asset-require-hook": "^1.2.0",
     "cleave.js": "^1.5.3",
     "compression": "^1.7.4",
     "cookie-parser": "^1.4.3",
-    "d3": "^5.9.7",
-    "deepmerge": "^4.0.0",
-    "dotenv": "^8.0.0",
+    "d3": "^5.15.0",
+    "deepmerge": "^4.2.2",
+    "dotenv": "^8.2.0",
     "es6-promise": "^4.2.8",
     "express": "^4.17.0",
     "express-useragent": "^1.0.13",
     "font-awesome": "^4.7.0",
-    "formdata-polyfill": "^3.0.18",
-    "getenv": "^0.7.0",
-    "gsap": "^3.0.1",
+    "formdata-polyfill": "^3.0.19",
+    "getenv": "^1.0.0",
+    "gsap": "^3.2.0",
     "http-link-header": "^1.0.2",
     "ignore-loader": "^0.1.2",
     "ignore-styles": "^5.0.1",
     "isomorphic-fetch": "^2.2.1",
+    "jsdom": "^16.2.0",
     "json-pointer": "^0.6.0",
     "lodash": "^4.17.15",
-    "mapbox-gl": "^1.5.0",
-    "markdown-to-jsx": "^6.10.2",
-    "mitt": "^1.1.3",
+    "mapbox-gl": "^1.8.0",
+    "markdown-to-jsx": "^6.11.0",
+    "mitt": "^1.2.0",
     "moment": "^2.24.0",
-    "node-polyglot": "^2.3.0",
+    "node-polyglot": "^2.4.0",
+    "nodemon": "^2.0.2",
     "normalize.css": "^8.0.1",
-    "path-to-regexp": "^3.0.0",
+    "path-to-regexp": "^6.1.0",
     "postcss-import": "^12.0.1",
     "postcss-preset-env": "^6.7.0",
     "prop-types": "^15.7.1",
-    "rc-tooltip": "^3.7.3",
-    "react": "^16.8.6",
-    "react-dom": "^16.8.6",
+    "react": "^16.13.0",
+    "react-dom": "^16.13.0",
     "react-faux-dom": "^4.5.0",
-    "react-joyride": "^2.1.1",
+    "react-joyride": "^2.2.1",
     "react-select": "^1.3.0",
     "react-simplemde-editor": "^4.1.0",
-    "react-tabs": "^3.0.0",
+    "react-tabs": "^3.1.0",
     "remove-markdown": "^0.3.0",
     "simplemde": "^1.11.2",
-    "source-sans-pro": "^2.45.0",
+    "source-sans-pro": "^3.6.0",
     "url-template": "^2.0.8",
-    "webpack": "^4.38.0",
-    "webpack-dev-middleware": "^3.7.0",
-    "webpack-dev-server": "^3.7.1",
+    "webpack": "^4.41.6",
+    "webpack-dev-middleware": "^3.7.2",
+    "webpack-dev-server": "^3.10.3",
     "webpack-hot-middleware": "^2.25.0"
   }
 }
