@@ -4,29 +4,32 @@ import I18nProvider from './I18nProvider'
 import EmittProvider from './EmittProvider'
 import ApiProvider from './ApiProvider'
 import UserProvider from './UserProvider'
+import ConfigProvider from './ConfigProvider'
 import App from './App'
 
 const Init = ({
-  i18n, emitter, children, apiConfig, supportedLanguages,
+  i18n, emitter, children, config, supportedLanguages,
 }) => (
-  <I18nProvider i18n={i18n}>
-    <EmittProvider emitter={emitter}>
-      <ApiProvider config={apiConfig}>
-        <UserProvider>
-          <App locales={i18n.locales} supportedLanguages={supportedLanguages}>
-            {children}
-          </App>
-        </UserProvider>
-      </ApiProvider>
-    </EmittProvider>
-  </I18nProvider>
+  <ConfigProvider config={config}>
+    <I18nProvider i18n={i18n}>
+      <EmittProvider emitter={emitter}>
+        <ApiProvider config={config.apiConfig}>
+          <UserProvider>
+            <App locales={i18n.locales} supportedLanguages={supportedLanguages}>
+              {children}
+            </App>
+          </UserProvider>
+        </ApiProvider>
+      </EmittProvider>
+    </I18nProvider>
+  </ConfigProvider>
 )
 
 Init.propTypes = {
   i18n: PropTypes.objectOf(PropTypes.any).isRequired,
   emitter: PropTypes.objectOf(PropTypes.any),
   children: PropTypes.node.isRequired,
-  apiConfig: PropTypes.objectOf(PropTypes.any).isRequired,
+  config: PropTypes.objectOf(PropTypes.any).isRequired,
   supportedLanguages: PropTypes.arrayOf(PropTypes.any).isRequired,
 }
 
