@@ -29,7 +29,7 @@ const Input = ({
         {
           __html: translate(title)
               + (required
-                ? `<span class="asterisk" title="${translate('This is a required field!')}">*</span>`
+                ? `<span class="asterisk" title="${translate('Error.requiredField')}">*</span>`
                 : ''),
         }}
       className={required ? 'required' : ''}
