@@ -45,7 +45,7 @@ class KeywordSelect extends React.Component {
         >
           {translate(title)}
           &nbsp;
-          {required ? <span className="asterisk" title={translate('This is a required field!')}>*</span> : ''}
+          {required ? <span className="asterisk" title={translate('Error.requiredField')}>*</span> : ''}
         </div>
         <span className="fieldDescription">
           {(description
