@@ -20,7 +20,7 @@ const List = ({
     >
       {translate(title)}
       &nbsp;
-      {required ? <span className="asterisk" title={translate('This is a required field!')}>*</span> : ''}
+      {required ? <span className="asterisk" title={translate('Error.requiredField')}>*</span> : ''}
     </div>
     <span className="fieldDescription">
       {(description
