@@ -15,7 +15,7 @@ const Select = ({
     >
       {translate(title)}
       &nbsp;
-      {required ? <span className="asterisk" title={`${translate('This is a required field!')}`}>*</span> : ''}
+      {required ? <span className="asterisk" title={`${translate('Error.requiredField')}`}>*</span> : ''}
     </label>
     <span className="fieldDescription">
       {(description
