@@ -25,9 +25,7 @@ const WebPage = ({
   feature,
   _links,
   _self,
-  mapboxConfig,
   schema,
-  embedValue,
   showOptionalFields,
   emitter,
   translate,
@@ -45,7 +43,7 @@ const WebPage = ({
         const modalDialog = document.querySelector('.WebPage')
         if (!modalDialog.contains(e.target)) {
           if (view === 'edit') {
-            confirm(translate('Do you want to go leave the edit view?')) && emitter.emit('navigate', _self || Link.home)
+            confirm(translate('WebPage.leaveEditView')) && emitter.emit('navigate', _self || Link.home)
           } else {
             emitter.emit('navigate', Link.home)
           }
@@ -62,7 +60,6 @@ const WebPage = ({
           view={view}
           _self={_self}
           _links={_links}
-          embedValue={embedValue}
         />
 
 
@@ -73,7 +70,6 @@ const WebPage = ({
               <WebPageEdit
                 about={about}
                 action={about['@id'] ? 'edit' : 'add'}
-                mapboxConfig={mapboxConfig}
                 schema={schema}
                 closeLink={about['@id'] ? _self : undefined}
                 showOptionalFields={showOptionalFields}
@@ -89,7 +85,6 @@ const WebPage = ({
               about={about}
               view={view}
               action={about['@id'] ? 'edit' : 'add'}
-              mapboxConfig={mapboxConfig}
               schema={schema}
               _self={_self}
               isLiveEvent={isLiveEvent}
@@ -112,15 +107,7 @@ WebPage.propTypes = {
   user: PropTypes.objectOf(PropTypes.any),
   _self: PropTypes.string.isRequired,
   _links: PropTypes.objectOf(PropTypes.any),
-  mapboxConfig: PropTypes.shape(
-    {
-      token: PropTypes.string,
-      style: PropTypes.string,
-      miniMapStyle: PropTypes.string,
-    },
-  ).isRequired,
   schema: PropTypes.objectOf(PropTypes.any).isRequired,
-  embedValue: PropTypes.string,
   showOptionalFields: PropTypes.bool,
   emitter: PropTypes.objectOf(PropTypes.any).isRequired,
   translate: PropTypes.func.isRequired,
@@ -133,7 +120,6 @@ WebPage.defaultProps = {
   contributor: null,
   dateModified: null,
   _links: { refs: [] },
-  embedValue: null,
   showOptionalFields: true,
   onSubmit: formdata => console.log(formdata),
 }
